requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
imageio>=2.31.0
tqdm>=4.66.0
lpips>=0.1.4
pytorch-fid>=0.3.0
requests>=2.31.0
gradio>=4.0.0
